Update GIT & Mercurial (hg) project VCS detection
The project root directory has a `.git` folder which the old prompt
detected and displayed in terminal. When you `cd` into a sub-folder the
prompt used to not detect GIT version control as there was no. `.git`
folder present in the sub-folder.

This same fix has been applied to  Mercurial version control, though it
has not been tested.